Provide dt (datetime) parameter for all bookmarks
When this field is not supplied, the Pinboard servers are currently
using the client's local time as though it were UTC. For example A
bookmark I create at 9am in EST (i.e. -05:00) is stored with a creation
date of 9am UTC (i.e. +00:00), which is off by 5 hours. Sending the dt
parameter results in the correct creation time being stored by Pinboard.